Mid-refactor for #17
