+++ afscgap/__init__.py
@@ -25,10 +25,12 @@ import typing
 
 import afscgap.client
 import afscgap.model
+import afscgap.query
 
 STR_OR_DICT = typing.Union[str, dict]
-FLOAT_PARAM = typing.Optional[typing.Union[float, dict]]
-INT_PARAM = typing.Optional[typing.Union[int, dict]]
+RANGE_TUPLE = typing.Tuple[float]
+FLOAT_PARAM = typing.Optional[typing.Union[float, dict, RANGE_TUPLE]]
+INT_PARAM = typing.Optional[typing.Union[int, dict, RANGE_TUPLE]]
 STR_PARAM = typing.Optional[STR_OR_DICT]
 
 
@@ -186,7 +188,7 @@ def query(
         Cursor to manage HTTP requests and query results.
     """
 
-    all_dict = {
+    all_dict_raw = {
         'year': year,
         'srvy': srvy,
         'survey': survey,
@@ -197,7 +199,7 @@ def query(
         'station': station,
         'vessel_name': vessel_name,
         'vessel_id': vessel_id,
-        'date_time': convert_from_iso8601(date_time),
+        'date_time': date_time,
         'latitude_dd': latitude_dd,
         'longitude_dd': longitude_dd,
         'species_code': species_code,
@@ -224,43 +226,4 @@ def query(
         'ak_survey_id': ak_survey_id
     }
 
-    query_url = afscgap.client.get_query_url(all_dict, base=base_url)
-    return afscgap.client.Cursor(
-        query_url,
-        limit=limit,
-        start_offset=start_offset,
-        requestor=requestor,
-        filter_incomplete=filter_incomplete
-    )
-
-
-def convert_from_iso8601(target: STR_PARAM) -> STR_PARAM:
-    """Convert strings from ISO 8601 format to API format.
-
-    Args:
-        target: The string or dictionary in which to perform the
-            transformations.
-
-    Returns:
-        If given an ISO 8601 string, will convert from ISO 8601 to the API
-        datetime string format. Similarly, if given a dictionary, all values
-        matching an ISO 8601 string will be converted to the API datetime string
-        format. If given None, returns None.
-    """
-    if target is None:
-        return None
-    elif isinstance(target, str):
-        return afscgap.model.convert_from_iso8601(target)
-    elif isinstance(target, dict):
-        items = target.items()
-        output_dict = {}
-
-        for key, value in items:
-            if isinstance(value, str):
-                output_dict[key] = afscgap.model.convert_from_iso8601(value)
-            else:
-                output_dict[key] = value
-
-        return output_dict
-    else:
-        return target
+    afscgap.client.build_cursor(all_dict_raw)
